Change Link.ruleId type from Int to Long

+++ app/src/main/java/me/alex/pet/apps/zhishi/domain/StyledText.kt
@@ -25,4 +25,4 @@ enum class CharacterStyleType {
     MISSPELL
 }
 
-data class Link(val start: Int, val end: Int, val ruleId: Int)
+data class Link(val start: Int, val end: Int, val ruleId: Long)
